requirements.txt
awscli==1.15.81
awsebcli==3.7.7
boto==2.41.0
botocore==1.10.80
colorama==0.3.7
Django==1.11
django-compressor==2.2
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
djangorestframework==3.7.0
